requirements.txt
aiosqlite
httpx
python-dotenv
python-telegram-bot
tenacity
psutil
